- add of int32; - Bug fix in the type for zin vector; - typing bug for the clock (init).

+++ lib/ztypes.ml
@@ -31,7 +31,7 @@ type ('a, 'b) node =
 type time = float
 type cvec = float array
 type dvec = float array
-type zinvec = float array
+type zinvec = int32 array
 type zoutvec = float array
 		    
 type ('a, 'b) hybrid =
